Rework the loop detection/handling
The level check from before would return before the manifest was copied.
This looks for loops in direct parents, and if encountered, it switches
to copy the referrers and digest tags after the manifest is pushed. The
manifest wait group is being removed to reduce complexity and remove
deadlock issues.

Signed-off-by: Brandon Mitchell <git@bmitch.net>