Fix regbot reproducible build
First, `SOURCE_DATE_EPOCH` was misspelled in the workflow file, so the
vcs_sec value did not get passed in correctly.

Second, the fallback had a syntax error:

```
    SOURCE_DATE_EPOCH="${SOURCE_DATE_EPOCH}:-$(date +%s)}"
                                          ^
```

The accidental `}` meant that something like `:-1696222361}` would
unconditionally get appended.  If `$SOURCE_DATE_EPOCH` was e.g.
`1234567890`, it would get changed to `1234567890:-1696222361}`.

If it wasn't set to begin with, it would become `:-1696222361}`.

Third, the reason it works when `SOURCE_DATE_EPOCH` is passed in
correctly is that the shell construct where `SOURCE_DATE_EPOCH` was to
be used does not work as expected:

    SOURCE_DATE_EPOCH="${SOURCE_DATE_EPOCH:-$(date +%s)}" find . -exec touch --date="@${SOURCE_DATE_EPOCH}" '{}' \;

Setting `SOURCE_DATE_EPOCH` in this manner causes it to be passed as an
environment variable to `find`. However, it is not `find` that expands
`@${SOURCE_DATE_EPOCH}` in the `--date` argument. That is done by the
shell. The shell has the original value of `SOURCE_DATE_EPOCH`.

Signed-off-by: Søren L. Hansen <soren.hansen@mastercard.com>