Feat: Add a priority queue for network requests
Breaking: Update scheme to use pqueue instead of throttle.

This is the first step to redesign the network requests for more efficiency.

Library users are unlikely to encounter issues with the scheme change since
both throttle and pqueue are internal packages, so it shouldn't be possible
to call their methods directly or specify their types in variables.

Signed-off-by: Brandon Mitchell <git@bmitch.net>