Feat: Support auth on redirect
A registry server that is just a redirect to a real server that requires auth will now work.
Previously, auth was only sent to the first server, now it is only sent to the server that performed the auth.
The auth package itself was also refactored a bit, removing unneeded interfaces and reducing exported fields.

Signed-off-by: Brandon Mitchell <git@bmitch.net>